requirements.txt
asgiref==3.7.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
drf-yasg==1.21.7
feedgenerator==2.1.0
filelock==3.12.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
openapi-codec==1.3.2
packaging==23.2
pelican==4.8.0
platformdirs==3.10.0
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.5
whitenoise==6.6.0